13. Оптимизируй это (часть 1)_корректировка_1
