+++ src/store/api-actions.js
@@ -84,7 +84,7 @@ const logout = () => (dispatch, _getState, api) => (
 const changeFavoriteStatus = (id, status) => (dispatch, _getState, api) => (
   api.post(`/favorite/${id}/${status}`, {id, status})
   .then(({data}) => {
-    dispatch(ActionCreator.setActiveCard(adaptPlaceCardToClient(data)));
+    dispatch(ActionCreator.changeFavoriteCard(adaptPlaceCardToClient(data)));
   })
 );
